feat: text input migration
