+++ docs/docs/guides/12-migration.md
@@ -1,0 +1,127 @@
+---
+title: Migration from Paper 5.x to 6.x
+---
+
+TBC
+
+## Component and types
+
+The Paper 6.x `TextInput` is a complete rewrite with a new API. Import the component the same way, but note that the props and behavior have changed significantly.
+
+```tsx
+import { TextInput, type TextInputProps } from 'react-native-paper';
+```
+
+## Visual / variant
+
+- **`mode="flat"`** → **`variant="filled"`**
+- **`mode="outlined"`** → **`variant="outlined"`**
+
+```tsx
+// Before (v5)
+<TextInput mode="flat" label="Filled" />
+<TextInput mode="outlined" label="Outlined" />
+
+// After (v6)
+<TextInput variant="filled" label="Filled" />
+<TextInput variant="outlined" label="Outlined" />
+```
+
+## Icons and adornments
+
+- **`left` / `right`** → **`startAccessory` / `endAccessory`**
+- **`TextInput.Icon`** → **`TextInput.Icon`**
+- **`TextInput.Affix`** → **`prefix` / `suffix`**, or **`TextInput.Icon`**, or **`startAccessory` / `endAccessory`**
+
+```tsx
+// Before (v5)
+<TextInput
+  left={<TextInput.Icon icon="email" />}
+  right={<TextInput.Affix text={`${value.length}/80`} />}
+/>
+
+// After (v6)
+<TextInput
+  startAccessory={(p) => <TextInput.Icon {...p} icon="email" />}
+  endAccessory={(p) => <CustomComponent {...p} />}
+  maxLength={100}
+  prefix="$"
+  suffix="/100"
+  counter
+/>
+```
+
+## Label, helper, error, disabled
+
+- **`label: React.Element | string`** → **`string`**
+- **`error` / `disabled`** → **`error`** and **`disabled`**
+- **`HelperText`** was removed; use **`supportingText`**.
+
+```tsx
+// Before (v5)
+<>
+  <TextInput
+    label="Email"
+    error={hasError}
+    disabled={isDisabled}
+  />
+  <HelperText type="error" visible={hasError}>
+    Enter a valid email
+  </HelperText>
+</>
+
+// After (v6)
+<TextInput
+  label="Email"
+  error={hasError}
+  disabled={isDisabled}
+  supportingText="Enter a valid email"
+/>
+```
+
+## Styling / behavior removed
+
+No direct `TextInput` equivalents for:
+
+- **`dense`**, **`contentStyle`**, **`underlineStyle`**
+- **`underlineColor`**, **`activeUnderlineColor`**, **`outlineColor`**, **`activeOutlineColor`**, **`textColor`**
+- **`render`**
+
+Use **`style`** on the inner input and the **`theme`** for colors.
+
+```tsx
+import { MD3LightTheme, TextInput } from 'react-native-paper';
+
+const theme = {
+  ...MD3LightTheme,
+  colors: {
+    ...MD3LightTheme.colors,
+    outline: '#79747E',
+    primary: '#6750A4',
+  },
+};
+
+// Before (v5)
+<TextInput
+  dense
+  contentStyle={{
+    paddingHorizontal: 12,
+    paddingTop: 8,
+    paddingBottom: 8,
+  }}
+  outlineStyle={{
+    borderRadius: 12,
+    borderWidth: 2,
+  }}
+  outlineColor="#79747E"
+  activeOutlineColor="#6750A4"
+  textColor="#1C1B1F"
+  style={{ fontSize: 16 }}
+/>
+
+// After (v6)
+<TextInput
+  theme={theme}
+  style={{ fontSize: 16, color: '#1C1B1F' }}
+/>
+```
