+++ docs/package.json
@@ -29,6 +29,7 @@
     "color": "^4.2.3",
     "marked": "^4.1.1",
     "prism-react-renderer": "^1.3.5",
+    "process": "^0.11.10",
     "react": "17.0.2",
     "react-color": "^2.19.3",
     "react-dom": "17.0.2",
