+++ docs/src/components/PropTable.tsx
@@ -11,10 +11,10 @@ const typeDefinitions = {
     'https://github.com/callstack/react-native-paper/blob/main/src/components/Icon.tsx#L16',
   ThemeProp:
     'https://callstack.github.io/react-native-paper/docs/guides/theming#theme-properties',
-  '(props: TextFieldAccessoryProps) => React.ReactNode':
-    'https://github.com/callstack/react-native-paper/blob/main/src/components/TextField/TextFieldIcon.tsx#L11',
-  '(props: TextFieldRenderProps) => React.ReactNode':
-    'https://github.com/callstack/react-native-paper/blob/main/src/components/TextField/TextField.tsx#L159',
+  '(props: TextInputAccessoryProps) => React.ReactNode':
+    'https://github.com/callstack/react-native-paper/blob/main/src/components/TextInput/TextInputIcon.tsx#L11',
+  '(props: TextInputRenderProps) => React.ReactNode':
+    'https://github.com/callstack/react-native-paper/blob/main/src/components/TextInput/TextInput.tsx#L168',
   AccessibilityState:
     'https://reactnative.dev/docs/accessibility#accessibilitystate',
   'StyleProp<ViewStyle>': 'https://reactnative.dev/docs/view-style-props',
