--- src/components/TextInput/Adornment/enums.tsx
@@ -1,12 +1,0 @@
-export enum AdornmentType {
-  Icon = 'icon',
-  Affix = 'affix',
-}
-export enum AdornmentSide {
-  Right = 'right',
-  Left = 'left',
-}
-export enum InputMode {
-  Outlined = 'outlined',
-  Flat = 'flat',
-}
