--- src/components/TextInput/Adornment/types.tsx
@@ -1,11 +1,0 @@
-import type { AdornmentSide, AdornmentType } from './enums';
-
-export type AdornmentConfig = {
-  side: AdornmentSide;
-  type: AdornmentType;
-};
-export type AdornmentStyleAdjustmentForNativeInput = {
-  adornmentStyleAdjustmentForNativeInput: Array<
-    { paddingRight: number; paddingLeft: number } | {}
-  >;
-};
