+++ src/components/TextInput/constants.ts
@@ -9,26 +9,26 @@ import { defaultShapes } from '../../theme/tokens/sys/shape';
 export const fontScale = PixelRatio.getFontScale();
 
 /**
- * Common constants for the text field component.
+ * Common constants for the text input component.
  */
 
-export const BASELINE_TEXT_FIELD_HEIGHT = 56;
-export const BASELINE_TEXT_FIELD_PADDING_VERTICAL = 8;
+export const BASELINE_TEXT_INPUT_HEIGHT = 56;
+export const BASELINE_TEXT_INPUT_PADDING_VERTICAL = 8;
 
-export const TEXT_FIELD_INPUT_WRAPPER_PADDING_HORIZONTAL = 16;
-export const TEXT_FIELD_ACCESSORY_MARGIN_HORIZONTAL = 12;
+export const TEXT_INPUT_INPUT_WRAPPER_PADDING_HORIZONTAL = 16;
+export const TEXT_INPUT_ACCESSORY_MARGIN_HORIZONTAL = 12;
 
-export const TEXT_FIELD_HEIGHT = Math.ceil(
-  BASELINE_TEXT_FIELD_HEIGHT * fontScale
+export const TEXT_INPUT_HEIGHT = Math.ceil(
+  BASELINE_TEXT_INPUT_HEIGHT * fontScale
 );
-export const TEXT_FIELD_PADDING_VERTICAL = Math.ceil(
-  BASELINE_TEXT_FIELD_PADDING_VERTICAL * fontScale
+export const TEXT_INPUT_PADDING_VERTICAL = Math.ceil(
+  BASELINE_TEXT_INPUT_PADDING_VERTICAL * fontScale
 );
 
-export const TEXT_FIELD_BORDER_RADIUS = defaultShapes.corner.extraSmall;
+export const TEXT_INPUT_BORDER_RADIUS = defaultShapes.corner.extraSmall;
 
 export const LABEL_START_OFFSET_WITHOUT_ACCESSORY =
-  TEXT_FIELD_INPUT_WRAPPER_PADDING_HORIZONTAL;
+  TEXT_INPUT_INPUT_WRAPPER_PADDING_HORIZONTAL;
 
 export const ACCESSORY_SIZE = 24;
 
@@ -64,27 +64,27 @@ const FILLED_LINE_HEIGHT_DELTA = 3;
 
 export const FILLED_LABEL_START_OFFSET_WITH_ACCESSORY =
   ACCESSORY_SIZE +
-  TEXT_FIELD_ACCESSORY_MARGIN_HORIZONTAL +
-  TEXT_FIELD_INPUT_WRAPPER_PADDING_HORIZONTAL;
+  TEXT_INPUT_ACCESSORY_MARGIN_HORIZONTAL +
+  TEXT_INPUT_INPUT_WRAPPER_PADDING_HORIZONTAL;
 
-export const FILLED_ACTIVE_LABEL_TOP_POSITION = TEXT_FIELD_PADDING_VERTICAL;
+export const FILLED_ACTIVE_LABEL_TOP_POSITION = TEXT_INPUT_PADDING_VERTICAL;
 
 export const FILLED_INACTIVE_LABEL_TOP_POSITION = Math.ceil(
-  ((BASELINE_TEXT_FIELD_HEIGHT -
-    2 * BASELINE_TEXT_FIELD_PADDING_VERTICAL -
+  ((BASELINE_TEXT_INPUT_HEIGHT -
+    2 * BASELINE_TEXT_INPUT_PADDING_VERTICAL -
     INPUT_FONT_SIZE) /
     2 +
-    BASELINE_TEXT_FIELD_PADDING_VERTICAL) *
+    BASELINE_TEXT_INPUT_PADDING_VERTICAL) *
     fontScale
 );
 
 export const FILLED_MULTILINE_PADDING_TOP =
-  Math.ceil(ACTIVE_LABEL_FONT_SIZE * fontScale) + TEXT_FIELD_PADDING_VERTICAL;
+  Math.ceil(ACTIVE_LABEL_FONT_SIZE * fontScale) + TEXT_INPUT_PADDING_VERTICAL;
 
 export const FILLED_DISABLED_CONTAINER_OPACITY = 0.04;
 
 export const FILLED_PADDING_BOTTOM =
-  TEXT_FIELD_PADDING_VERTICAL + FILLED_LINE_HEIGHT_DELTA;
+  TEXT_INPUT_PADDING_VERTICAL + FILLED_LINE_HEIGHT_DELTA;
 
 /**
  * Constants for the outlined variant.
@@ -95,8 +95,8 @@ const OUTLINED_LINE_HEIGHT_DELTA = 2;
 export const OUTLINED_DISABLED_OUTLINE_OPACITY = 0.12;
 
 export const OUTLINED_MULTILINE_PADDING_TOP = Math.ceil(
-  ((BASELINE_TEXT_FIELD_HEIGHT -
-    2 * BASELINE_TEXT_FIELD_PADDING_VERTICAL -
+  ((BASELINE_TEXT_INPUT_HEIGHT -
+    2 * BASELINE_TEXT_INPUT_PADDING_VERTICAL -
     INPUT_FONT_SIZE) /
     2 -
     OUTLINED_LINE_HEIGHT_DELTA) *
@@ -107,21 +107,21 @@ export const OUTLINED_LABEL_PADDING_HORIZONTAL = 4;
 
 export const OUTLINED_LABEL_START_OFFSET_WITH_ACCESSORY =
   ACCESSORY_SIZE +
-  TEXT_FIELD_ACCESSORY_MARGIN_HORIZONTAL +
-  TEXT_FIELD_INPUT_WRAPPER_PADDING_HORIZONTAL -
+  TEXT_INPUT_ACCESSORY_MARGIN_HORIZONTAL +
+  TEXT_INPUT_INPUT_WRAPPER_PADDING_HORIZONTAL -
   OUTLINED_LABEL_PADDING_HORIZONTAL;
 
 export const OUTLINED_ACTIVE_LABEL_TOP_POSITION = Math.ceil(
-  (-BASELINE_TEXT_FIELD_PADDING_VERTICAL + OUTLINED_LINE_HEIGHT_DELTA) *
+  (-BASELINE_TEXT_INPUT_PADDING_VERTICAL + OUTLINED_LINE_HEIGHT_DELTA) *
     fontScale
 );
 
 export const OUTLINED_INACTIVE_LABEL_TOP_POSITION = Math.ceil(
-  ((BASELINE_TEXT_FIELD_HEIGHT -
-    2 * BASELINE_TEXT_FIELD_PADDING_VERTICAL -
+  ((BASELINE_TEXT_INPUT_HEIGHT -
+    2 * BASELINE_TEXT_INPUT_PADDING_VERTICAL -
     INPUT_FONT_SIZE) /
     2 +
-    BASELINE_TEXT_FIELD_PADDING_VERTICAL -
+    BASELINE_TEXT_INPUT_PADDING_VERTICAL -
     OUTLINED_LINE_HEIGHT_DELTA) *
     fontScale
 );
@@ -129,7 +129,7 @@ export const OUTLINED_INACTIVE_LABEL_TOP_POSITION = Math.ceil(
 /** Positive distance; apply sign in animation using `isRTL` from `useLocale`. */
 export const OUTLINED_LABEL_TRANSLATE_DISTANCE_WITH_ACCESSORY =
   ACCESSORY_SIZE +
-  TEXT_FIELD_INPUT_WRAPPER_PADDING_HORIZONTAL -
+  TEXT_INPUT_INPUT_WRAPPER_PADDING_HORIZONTAL -
   OUTLINED_LABEL_PADDING_HORIZONTAL;
 
 /** Positive distance; apply sign in animation using `isRTL` from `useLocale`. */
