+++ src/components/TextInput/index.ts
@@ -1,0 +1,13 @@
+import TextInputComponent from './TextInput';
+import TextInputIcon from './TextInputIcon';
+
+const TextInput = Object.assign(
+  // @component ./TextInput.tsx
+  TextInputComponent,
+  {
+    // @component ./TextInputIcon.tsx
+    Icon: TextInputIcon,
+  }
+);
+
+export default TextInput;
