+++ src/index.tsx
@@ -37,7 +37,6 @@ export { default as Dialog } from './components/Dialog/Dialog';
 export { default as Divider } from './components/Divider';
 export { default as FAB } from './components/FAB';
 export { default as AnimatedFAB } from './components/FAB/AnimatedFAB';
-export { default as HelperText } from './components/HelperText/HelperText';
 export { default as Icon } from './components/Icon';
 export { default as IconButton } from './components/IconButton/IconButton';
 export { default as Menu } from './components/Menu/Menu';
@@ -51,8 +50,7 @@ export { default as Surface } from './components/Surface';
 export { default as Switch } from './components/Switch/Switch';
 export { default as Appbar } from './components/Appbar';
 export { default as TouchableRipple } from './components/TouchableRipple/TouchableRipple';
-export { default as TextInput } from './components/TextInput/TextInput';
-export { default as TextField } from './components/TextField';
+export { default as TextInput } from './components/TextInput';
 export { default as ToggleButton } from './components/ToggleButton';
 export { default as SegmentedButtons } from './components/SegmentedButtons/SegmentedButtons';
 export { default as Tooltip } from './components/Tooltip/Tooltip';
@@ -105,7 +103,6 @@ export type { Props as DrawerItemProps } from './components/Drawer/DrawerItem';
 export type { Props as DrawerSectionProps } from './components/Drawer/DrawerSection';
 export type { Props as FABProps } from './components/FAB/FAB';
 export type { Props as FABGroupProps } from './components/FAB/FABGroup';
-export type { Props as HelperTextProps } from './components/HelperText/HelperText';
 export type { Props as IconButtonProps } from './components/IconButton/IconButton';
 export type { Props as ListAccordionProps } from './components/List/ListAccordion';
 export type { Props as ListAccordionGroupProps } from './components/List/ListAccordionGroup';
@@ -129,19 +126,16 @@ export type { Props as SearchbarProps } from './components/Searchbar';
 export type { Props as SnackbarProps } from './components/Snackbar';
 export type { Props as SurfaceProps } from './components/Surface';
 export type { Props as SwitchProps } from './components/Switch/Switch';
-export type { Props as TextInputProps } from './components/TextInput/TextInput';
-export type { Props as TextInputAffixProps } from './components/TextInput/Adornment/TextInputAffix';
-export type { Props as TextInputIconProps } from './components/TextInput/Adornment/TextInputIcon';
 export type {
-  TextFieldProps,
-  TextFieldRenderProps,
-  TextFieldVariant,
-  TextFieldHandles,
-} from './components/TextField/TextField';
+  TextInputProps,
+  TextInputRenderProps,
+  TextInputVariant,
+  TextInputHandles,
+} from './components/TextInput/TextInput';
 export type {
-  TextFieldAccessoryProps,
-  TextFieldIconProps,
-} from './components/TextField/TextFieldIcon';
+  TextInputAccessoryProps,
+  TextInputIconProps,
+} from './components/TextInput/TextInputIcon';
 export type { Props as ToggleButtonProps } from './components/ToggleButton/ToggleButton';
 export type { Props as ToggleButtonGroupProps } from './components/ToggleButton/ToggleButtonGroup';
 export type { Props as ToggleButtonRowProps } from './components/ToggleButton/ToggleButtonRow';
